added .clangd and compile_commands.json for LSP
